+++ swift-sdk/IterableSDK.h
@@ -14,4 +14,5 @@ FOUNDATION_EXPORT double swift_sdkVersionNumber;
 //! Project version string for swift_sdk.
 FOUNDATION_EXPORT const unsigned char swift_sdkVersionString[];
 
-// In this header, you should import all the public headers of your framework using statements like #import <IterableSDK/IterableConstants.h>
+// In this header, you should import all the public headers of your framework using statements like #import <swift_sdk/PublicHeader.h>
+